requirements.txt
joker
pyyaml~=5.3.1
volkanic~=0.2.1
setuptools~=49.2.1